Remove PoolSize setting in Redis connection
The PoolSize setting has been removed from the code in the Redis connection settings. This was an integral part of using the NATS proxy connection pool, but it has been clarified that it can lead to errors. This change is aimed at ensuring more stable connections while interacting with the Redis client.